gateway/sap_qm_mock/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
aiosqlite>=0.20.0
paho-mqtt>=2.0
pydantic>=2.0
